add authorization

+++ controllers/cards.js
@@ -1,15 +1,17 @@
 const mongoose = require('mongoose');
-const http2 = require('node:http2');
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const BadRequest = require('../errors/bad-request');
+const ForbiddenError = require('../errors/forbiddenError');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate('owner')
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(http2.constants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию' }));
+    .catch((err) => next(err));
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user;
 
@@ -17,34 +19,42 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err instanceof mongoose.Error.ValidationError) {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки.' });
+        next(new BadRequest('Переданы некорректные данные при создании карточки'));
       } else {
-        res.status(http2.constants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
+  const userId = req.user._id;
 
-  Card.findByIdAndRemove({ _id: cardId })
-    .then((card) => {
-      if (card) {
-        res.send({ data: cardId });
+  Card.findById({ _id: cardId })
+    .then((cardInfo) => {
+      if (cardInfo) {
+        if (cardInfo.owner._id.toString() === userId) {
+          Card.findByIdAndRemove({ _id: cardId })
+            .then((card) => {
+              if (card) {
+                res.send({ data: cardId });
+              }
+            });
+        } else { next(new ForbiddenError('Вы не создатель карточки')); }
       } else {
-        res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена.' });
+        throw new NotFoundError('Карточка с указанным _id не найдена.');
       }
     })
     .catch((err) => {
       if (err instanceof mongoose.CastError) {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Не верный _id карточки' });
+        next(new BadRequest('Переданы некоректные данные при создании карточки'));
       } else {
-        res.status(http2.constants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -54,18 +64,18 @@ module.exports.likeCard = (req, res) => {
     .then((card) => {
       if (card) {
         res.send({ data: card });
-      } else { res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' }); }
+      } else { throw new NotFoundError('Передан несуществующий _id карточки'); }
     })
     .catch((err) => {
       if (err instanceof mongoose.CastError) {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки / снятия лайка' });
+        next(new BadRequest('Передан некорректный _id карточки'));
       } else {
-        res.status(http2.constants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -75,13 +85,13 @@ module.exports.dislikeCard = (req, res) => {
     .then((card) => {
       if (card) {
         res.send({ data: card });
-      } else { res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки' }); }
+      } else { throw new NotFoundError('Передан несуществующий _id карточки'); }
     })
     .catch((err) => {
       if (err instanceof mongoose.CastError) {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки / снятия лайка' });
+        next(new BadRequest('Переданы некорректные данные при создании карточки'));
       } else {
-        res.status(http2.constants.HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию' });
+        next(err);
       }
     });
 };
